Python Avengers/ScorePy-main/requirements.txt
pandas
numpy
matplotlib
seaborn
Pillow
customtkinter
